pyproject.toml
[project]
name = "etools"
version = "12"
description = "Default template for PDM package"

authors = [
    {name = "Robert Avram", email = "robert@ravdev.com"},
    {name = "Nikola Trncic", email = "ntrncic@unicef.org"},
]

dependencies = [
    'azure-common==1.1.28',
    'azure-storage-blob==12.20.0',
    'azure-storage-common==2.1.0',
    'carto==1.11.3',
    'celery==5.4.0',
    'cryptography==42.0.7',
    'dj-database-url==2.1.0',
    'dj-static==0.0.6',
    'Django==4.2.15',
    'django-admin-extra-urls==4.1.1',
    'django-appconf==1.0.6',
    'django-celery-beat==2.6.0',
    'django-celery-email==3.0.0',
    'django-celery-results==2.5.1',
    'django-contrib-comments==2.2.0',
    'django-cors-headers==4.3.1',
    'django-easy-pdf3==0.1.4',
    'django-extensions==3.2.3',
    'django-filter==24.2',
    'django-fsm==3.0.0',
    'django-import-export==4.0.2',
    'django-js-asset==2.2.0',
    'django-leaflet==0.30.0',
    'django-logentry-admin==1.1.0',
    'django-model-utils==4.5.1',
    'django-ordered-model==3.7.4',
    'django-post-office==3.8.0',
    'django-redis-cache==3.0.1',
    'django-rest-swagger==2.2.0',
    "django-storages[azure]==1.13.2",
    'django-tenants==3.6.1',
    'django-timezone-field==6.1.0',
    'django-waffle==4.1.0',
    'djangorestframework==3.15.1',
    'djangorestframework-csv==3.0.2',
    'djangorestframework-gis==1.0',
    'djangorestframework-recursive==0.1.2',
    'djangorestframework-simplejwt==5.3.1',
    'djangorestframework-xml==2.0.0',
    'drf-nested-routers==0.94.1',
    'drf-querystringfilter==1.0.0',
    'etools-offline==0.1.0',
    'etools-validator==0.5.1',
    'flower==2.0.1',
    'gunicorn==22.0.0',
    'msal==1.28.0',
    'newrelic==9.9.1',
    'openpyxl==3.1.2',
    'psycopg2-binary==2.9.9',
    'pyyaml==6.0.1',
    'reportlab==4.0.9',
    'requests==2.31.0',
    "sentry-sdk==1.27.0",
    'social-auth-app-django==5.4.1',
    "social-auth-core[azuread]==4.5.4",
    'tenant-schemas-celery==2.2.0',
    'unicef-attachments==0.12',
    'unicef-djangolib==0.7',
    'unicef-locations==4.2',
    "unicef-notification==1.4.1",
    'unicef-rest-export==0.6',
    'unicef-restlib==0.7',
    'unicef-snapshot==1.3',
    'unicef-vision==0.6',
    "gdal==3.8.5",
    "jsonschema>=4.22.0",
    "setuptools>=70.1.1",
    "django-debug-toolbar==4.2.0",
    "tblib>=3.0.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[tool.pdm.dev-dependencies]
dev = [
    "flake8",
    "coverage",
    "mock",
    "freezegun",
    "responses",
    "isort",
    "ipython",
    "pdbpp",
    "tox",
    "factory-boy",
    "sphinx",
    "vcrpy",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
